created file for test helper functions, removed unnecessary allThrough
